requirements.txt
Flask==2.3.3
torch==2.1.0
torchvision==0.16.0
Pillow==10.0.1
numpy==1.24.3
opencv-python==4.8.1.78
moviepy==1.0.3
ffmpeg-python==0.2.0
